fix: exhaustive production audit — correctness, security, performance, and architecture
Comprehensive codebase audit addressing 100 findings across all layers:

CRITICAL FIXES:
- Fix deadlock in MassiveOptionGateway.clear_caches() (Lock to RLock)
- Fix CAS commit ordering in scan/sweep services (rollback orphaned results)
- Fix error sanitization bypass (patterns before truncation, regex widened)
- Fix async get_market_holidays 404 parity with sync client
- Fix backtest quota off-by-one (exclude current job from count)

DECIMAL PRECISION:
- Convert all engine P&L math from float to Decimal
- TradeResult, EquityPointResult, PositionSnapshot financial fields to Decimal
- Eliminates cumulative IEEE 754 drift across hundreds of trades

TRANSACTIONAL OUTBOX:
- Complete the outbox pattern: job stays queued when inline send fails
- poll_outbox recovers pending messages with retry counting (max 10)
- Correlated job marked failed after outbox retries exhausted

STRIPE CLEANUP RETRY:
- New cleanup_stripe_orphan Celery task with 5 retries + exponential backoff
- Dispatched when synchronous cleanup fails during account deletion

SSE ISOLATION:
- Per-connection Redis clients for Pub/Sub subscriptions
- One connection failure no longer disrupts all SSE streams

MASSIVE CLIENT DEDUP:
- Extract _MassiveClientCore base class with shared config/parsing
- 944 to 665 lines (30% reduction)

FEATURE FLAGS:
- New feature_flags.py with percentage rollouts and tier targeting
- Deterministic CRC32 hashing for consistent user experience

SECURITY:
- Auto-generate random ip_hash_salt in dev
- Referer check as CSRF defense-in-depth for cookie auth
- Sec-Fetch-Site cross-site detection
- DELETE method added to chunked body limit check

PERFORMANCE:
- Bulk INSERT for trades/equity points
- GDPR export: 12 queries to 7
- Redis scan_iter to O(1) per-symbol metadata hash
- Separate celery_result_backend_url config
- CSV export: StringIO to BytesIO with size tracking

SCHEMA/MIGRATION:
- Add updated_at trigger for daily_recommendations
- Merge orphaned 0024_heartbeat migration branch
- Fix broken down_revision reference
- AuditEvent user_id=None on account deletion

FRONTEND-BACKEND CONTRACTS:
- ErrorEnvelope: add detail and details fields
- API version single source of truth
- OpenAPI snapshot and TypeScript types regenerated

TESTING (10 new test files, 57 tests):
- CAS rollback, Sortino consistency, sanitization bypass
- Deadlock prevention, Decimal precision, sweep limits
- Stripe retry, async parity, outbox recovery, contract parity
- Feature flag rollout behavior

Made-with: Cursor

+++ src/backtestforecast/market_data/types.py
@@ -12,7 +12,7 @@ class DailyBar:
     high_price: float
     low_price: float
     close_price: float
-    volume: int
+    volume: float
 
 
 @dataclass(frozen=True, slots=True)
@@ -21,7 +21,7 @@ class OptionContractRecord:
     contract_type: str
     expiration_date: date
     strike_price: float  # Precision note: float cannot represent all decimals exactly. Use Decimal(str(strike_price)) at financial computation boundaries.
-    shares_per_contract: int
+    shares_per_contract: float
 
 
 @dataclass(frozen=True, slots=True)
